// types/types.d.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** String representation of a Buffer ID. */
  Binary: any;
  DateTime: any;
  /** A field whose value conforms to the standard internet email address format as specified in RFC822: https://www.w3.org/Protocols/rfc822/. */
  EmailAddress: any;
};
// ...
export type CreateProductInput = {
  description: Scalars['String'];
  name: Scalars['String'];
};